fixes and development

+++ src/main/java/com/app/service/impl/ManufacturerServiceImpl.java
@@ -1,23 +1,24 @@
 package com.app.service.impl;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import com.app.entity.Manufacturer;
 import com.app.repository.ManufacturerRepository;
 import com.app.service.ManufacturerService;
 import com.app.utils.Result;
 import com.app.validator.ManufacturerValidator;
+import org.springframework.stereotype.Service;
 
 @Service("manufacturerService")
 public class ManufacturerServiceImpl implements ManufacturerService {
 	
-	@Autowired
-	private ManufacturerRepository manufacturerRepository;
+	private final ManufacturerRepository manufacturerRepository;
 	
-	@Autowired
-	private ManufacturerValidator manufacturerValidator;
-	
+	private final ManufacturerValidator manufacturerValidator;
+
+	public ManufacturerServiceImpl(ManufacturerRepository manufacturerRepository, ManufacturerValidator manufacturerValidator) {
+		this.manufacturerRepository = manufacturerRepository;
+		this.manufacturerValidator = manufacturerValidator;
+	}
+
 	public Result saveManufacturer(Manufacturer manufacturer) {
 		Result result = manufacturerValidator.checkBeforeSave(manufacturer);
 		
@@ -27,8 +28,13 @@ public class ManufacturerServiceImpl implements ManufacturerService {
 		return result;
 	}
 	
-	public void deleteManufacturer(Manufacturer manufacturer) {
-		manufacturerRepository.delete(manufacturer);
+	public Result deleteManufacturer(Manufacturer manufacturer) {
+		Result result = manufacturerValidator.checkBeforeDelete(manufacturer);
+
+		if (result.isSuccess())
+			manufacturerRepository.delete(manufacturer);
+
+		return result;
 	}
